+++ modules/cssmapmodule/config/app.php
@@ -14,9 +14,9 @@
  */
 return [
     'modules' => [
-        'site-module' => [
-            'class' => \modules\sitemodule\SiteModule::class,
+        'cssmap-module' => [
+            'class' => \modules\cssmapmodule\CSSMapModule::class,
         ],
     ],
-    'bootstrap' => ['site-module'],
+    'bootstrap' => ['cssmap-module'],
 ];