+++ modules/cssmapmodule/src/translations/en/cssmap-module.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * CSSMap module for Craft CMS 3.x
+ *
+ * Map css names to uglified names.
+ *
+ * @link      https://pennebaker.com
+ * @copyright Copyright (c) 2018 Pennebaker
+ */
+
+/**
+ * CSSMap en Translation
+ *
+ * Returns an array with the string to be translated (as passed to `Craft::t('cssmap-module', '...')`) as
+ * the key, and the translation as the value.
+ *
+ * http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html
+ *
+ * @author    Pennebaker
+ * @package   CSSMapModule
+ * @since     1.0.0
+ */
+return [
+    'CSSMap plugin loaded' => 'CSSMap plugin loaded',
+];