+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.0.21] - 2018.02.15
+* Updated babel to latest version
+* Implement CSS Map module which is a useful twig filter, when uglifying the CSS to id and class names.
+
 ## [1.0.20] - 2018.02.15
 ### Added
 * Added load function to `functions.es6`